requirements.txt
nats-py
pyyaml